Added an alpha setting to Components with Component::setAlpha(). Removed Component::fadeOutComponent method and instead refactored the ComponentAnimator class to provide fading in and out of the alpha level. Added a method Desktop::getAnimator() that provides a global instance of a ComponentAnimator so that there's no need to create your own.

+++ src/gui/graphics/drawables/juce_DrawableRectangle.cpp
@@ -73,8 +73,8 @@ bool DrawableRectangle::rebuildPath (Path& path) const
     const float w = Line<float> (points[0], points[1]).getLength();
     const float h = Line<float> (points[0], points[2]).getLength();
 
-    const float cornerSizeX = cornerSize.x.resolve (parent);
-    const float cornerSizeY = cornerSize.y.resolve (parent);
+    const float cornerSizeX = (float) cornerSize.x.resolve (parent);
+    const float cornerSizeY = (float) cornerSize.y.resolve (parent);
 
     path.clear();
